[wip] channel id fun

+++ cpp/foxglove/include/foxglove/server.hpp
@@ -106,12 +106,12 @@ struct WebSocketServerCallbacks {
   ///
   /// Only invoked if the channel is associated with the server and isn't already subscribed to by
   /// the client.
-  std::function<void(uint64_t channel_id, const ClientMetadata& client_metadata)> onSubscribe;
+  std::function<void(uint32_t channel_id, const ClientMetadata& client_metadata)> onSubscribe;
 
   /// @brief Callback invoked when a client unsubscribes from a channel.
   ///
   /// Only invoked for channels that had an active subscription from the client.
-  std::function<void(uint64_t channel_id, const ClientMetadata& client_metadata)> onUnsubscribe;
+  std::function<void(uint32_t channel_id, const ClientMetadata& client_metadata)> onUnsubscribe;
 
   /// @brief Callback invoked when a client advertises a client channel.
   ///
